+++ src/pretalx/api/serializers/schedule.py
@@ -86,7 +86,6 @@ class TalkSlotSerializer(FlexFieldsSerializerMixin, PretalxSerializer):
             "schedule",
             "description",
             "duration",
-            "board_number",
         ]
         read_only_fields = ["submission", "schedule"]
         expandable_fields = {