+++ src/pretalx/schedule/interfaces/widget.py
@@ -64,7 +64,6 @@ def build_widget_data(
                 "start": talk.local_start,
                 "end": talk.local_end,
                 "room": talk.room_id,
-                "board_number": talk.board_number,
                 "duration": talk.submission.get_duration(),
                 "updated": talk.updated.isoformat(),
                 "content_locale": talk.submission.content_locale,